Double the size of the Adler checksum
I'm committing this for archival purposes, it was an experiment that was supposed to figure out if it was possible to speed up the memcpy by adding more checksums to write to. My idea was that the current implementation is bottlenecked by the constant writes to the `b` part of the ckecksum that depend on the result of the writes to `a`, however that was not the case.
It's really difficult to profile this properly. Perf reports that there are no frontend stalls, I can't tell if there are any pipeline stalls, and callgrind is too slow to saturate the bandwidth for a more in-depth analysis.
Also the AdlerMemcpyAsm is entirely missing because I have no clue how to write ASM by hand, and using intrinsics to recreate the function won't work on clang until > 14 is released, as apparently I found a bug in the LLVM optimization pipeline that removes nontemporal hints from loads.